+++ NEWS.md
@@ -7,6 +7,8 @@
     - `clean_theme()`: remove axis lines, ticks, texts and titles.
     - `get_legend()`: to extract the legend labels from a ggplot object.
     - `as_ggplot()`: Transform the output of `gridExtra::arrangeGrob()` and `gridExtra::grid.arrange()` to a an object of class ggplot.
+    - `ggtexttable()`: to draw a textual table.
+      
       
 ## Major changes